requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
httpx==0.24.1
SQLAlchemy==2.0.23
aiohttp==3.9.1
Pillow==10.1.0 